requirements.txt
holehe
socialscan
sherlock-project
phonenumbers
PyPDF2
python-docx
openpyxl
python-pptx
mutagen
exifread
Pillow
pyheif
rarfile
hachoir
piexif
requests
hashid
rich
sqlmap
hashid
